ContractsPage Refactor
