+++ src/custom.scss
@@ -10,10 +10,6 @@ $info: #4800ff;
   max-width: 1200px;
 }
 
-.mw-600 {
-  max-width: 600px;
-}
-
 .mw-350 {
   width: 350px;
 }
